refactor: extract config-update route handler from smart-server
